refactor openledger
