+++ src/modules/openledger/actions.js
@@ -1,14 +1,5 @@
-import PersistentStorage from '../services/persistent-storage';
-import API from '../services/api';
-import * as types from '../mutations';
-
-const initialState = {
-  depositAddress: '',
-  pending: false,
-  coins: false,
-  error: false,
-  pendingAddress: false
-};
+import API from '../../services/api';
+import { types } from './mutations';
 
 const actions = {
   async checkIfAddressIsValid(state, { asset, address }) {
@@ -16,13 +7,12 @@ const actions = {
     return valid;
   },
   async fetchDepositAddress(store, { asset }) {
-    const { commit, rootGetters } = store;
-    const user = rootGetters['account/getCurrentUserName'];
+    const { commit, rootGetters, getters } = store;
+    const user = rootGetters['acc/getCurrentUserName'];
 
     commit(types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_REQUEST);
 
-    const cachedAddresses = PersistentStorage.getOpenledgerAddresses();
-
+    const cachedAddresses = getters.getDepositAddress;
     if (cachedAddresses[asset]) {
       const address = cachedAddresses[asset];
       commit(types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_COMPLETE, { address });
@@ -32,11 +22,9 @@ const actions = {
         asset,
         user
       });
-
       if (lastAddress.success) {
         const address = lastAddress.data;
         cachedAddresses[asset] = address;
-        PersistentStorage.setOpenledgerAddresses(cachedAddresses);
         commit(types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_COMPLETE, { address });
       } else {
         const newAddress = await API.Openledger.requestDepositAddress({
@@ -47,7 +35,6 @@ const actions = {
         if (newAddress.success) {
           const address = newAddress.data;
           cachedAddresses[asset] = address;
-          PersistentStorage.setOpenledgerAddresses(cachedAddresses);
           commit(types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_COMPLETE, { address });
         } else {
           const { error } = lastAddress;
@@ -83,44 +70,4 @@ const actions = {
   }
 };
 
-const getters = {
-  getDepositAddress: state => state.depositAddress,
-  getCoinsData: state => state.coins,
-  getAddressPending: state => state.pendingAddress
-};
-
-const mutations = {
-  [types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_REQUEST]: (state) => {
-    state.pendingAddress = true;
-    state.depositAddress = initialState.depositAddress;
-    state.error = null;
-  },
-  [types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_COMPLETE]: (state, { address }) => {
-    state.depositAddress = address;
-    state.pendingAddress = false;
-  },
-  [types.FETCH_OPENLEDGER_DEPOSIT_ADDRESS_ERROR]: (state, { error }) => {
-    state.error = error;
-    state.depositAddress = null;
-    state.pendingAddress = false;
-  },
-  [types.FETCH_OPENLEDGER_COINS_REQUEST]: (state) => {
-    state.pending = true;
-  },
-  [types.FETCH_OPENLEDGER_COINS_COMPLETE]: (state, { coins }) => {
-    state.pending = false;
-    state.coins = coins;
-  },
-  [types.FETCH_OPENLEDGER_COINS_ERROR]: (state, { error }) => {
-    state.pending = false;
-    state.error = error;
-  }
-};
-
-export default {
-  state: initialState,
-  actions,
-  getters,
-  mutations,
-  namespaced: true
-};
+export default actions;
