+++ src/modules/openledger/defaultState.js
@@ -1,0 +1,9 @@
+export const getDefaultState = () => {
+  return {
+    depositAddress: {},
+    pending: false,
+    coins: false,
+    error: false,
+    pendingAddress: false
+  };
+};
