+++ src/modules/openledger/getters.js
@@ -1,0 +1,7 @@
+const getters = {
+  getDepositAddress: state => state.depositAddress,
+  getCoinsData: state => state.coins,
+  getAddressPending: state => state.pendingAddress
+};
+
+export default getters;
